+++ Engine/Core/FileLoader.hpp
@@ -1,8 +1,9 @@
 #pragma once
 
 #include <string>
+#include <Core/VarTypes/Base.hpp>
 
 namespace VSGE {
-	bool LoadFile(const std::string& FilePath, char** out, size_t *Size = nullptr, bool Terminate = false);
-	bool LoadFilePartially(const std::string& FilePath, char** out, size_t Offset, size_t Size);
+	bool LoadFile(const std::string& FilePath, char** out, uint32 *Size = nullptr, bool Terminate = false);
+	bool LoadFilePartially(const std::string& FilePath, char** out, uint32 Offset, uint32 Size);
 }
